requirements.txt
binance-connector~=2.0.0
keras~=2.12.0
numpy~=1.24.2
pandas~=1.5.3
requests~=2.28.2
scikit-learn~=1.2.2
ta~=0.10.2
tensorflow~=2.12.0
websocket~=0.2.1

tqdm~=4.65.0
ccxt~=3.0.41